Improve settings control
Range and rate setters now accept the predefined values, and are overall optimized for this use.
Changed default settings in begin() to leave the ranges alone, and only set the rates to the minimum values just to turn them on.
Changed default conversion factors to match new range defaults
Removed unused member variables and methods. This is currently just a simplified port of the Arduino library, these can be implemented later if desired.